// frontend/src/types/user.ts
export interface User {
  address: string;
  balance: number;
  lastActivity: string;
  // Add other fields wey una need for user profile
  //   isOnline: boolean;
  //   profilePicture?: string; // Optional field for user profile picture
  //   username?: string; // Optional field for user display name
}